Telegram bot vl25 v2

+++ TelegramBot/RoomMaster.py
@@ -8,24 +8,36 @@ import asyncio
 # Собственно класс комнаты
 class Room:
 
-    def __init__(self, id_in):
-        self.room_id = id_in      # ID комнаты
-        self.cycle_list = []      # Список ТГ-ИД игроков в порядке их обхода (тот самый "цикл")
-        self.player_map = {}      # ТГ-ИД --> Player словарь
-        self.cycle_map = {}       # ТГ-ИД --> ind словарь, где ind - индекс чела в массива cycle_list
-        self.timer = None         # Текущий таймер
-        self.stage = 0            # Стадия/режим работы: 0 == игра не начата, # 1 == игроки описывают картинку,
-                                  # 2 == нейронка генерит картинки
-        self.round = -1           # Номер раунда (только текстовые!)
-        self.max_rounds = -1      # Кол-во раундов (только текстовые!)
-        self.tasks = []           # Пока что не используется
-        self.image_history = {}   # ТГ-ИД --> [картинка], хранит цепочки картинок для альбома
-        self.text_history = {}    # ТГ-ИД --> [текст], хранит цепочки подписей к картинкам для альбома
+    def __init__(self, id_in, chat_id):
+        self.room_id = id_in            # ID комнаты
+        self.cycle_list = []            # Список ТГ-ИД игроков в порядке их обхода (тот самый "цикл")
+        self.player_map = {}            # ТГ-ИД --> Player словарь
+        self.cycle_map = {}             # ТГ-ИД --> ind словарь, где ind - индекс чела в массива cycle_list
+        self.timer = None               # Текущий таймер
+        self.stage = 0                  # Стадия/режим работы: 0 == игра не начата, # 1 == игроки описывают картинку,
+                                        # 2 == нейронка генерит картинки
+        self.round = -1                 # Номер раунда (только текстовые!)
+        self.max_rounds = -1            # Кол-во раундов (только текстовые!)
+        self.tasks = []                 # Пока что не используется
+        self.image_history = {}         # ТГ-ИД --> [картинка], хранит цепочки картинок для альбома
+        self.text_history = {}          # ТГ-ИД --> [текст], хранит цепочки подписей к картинкам для альбома
+        self.time_for_round = 30        # времени на один раунд (сек)
+        self.generation_quality = 5     # качество генерации картинок
+        self.chat_id = chat_id          # ИД чата, в котором бот
 
     # отправить сообщение всем игрокам
-    async def send_plain_all(self, message):
-        for user_id in self.player_map.copy().keys():
-            await BotAPI.send_plain_text(user_id, message)
+    async def send_plain_all(self, message, send_to_group=False):
+        if send_to_group and self.chat_id != 0:
+            await BotAPI.send_plain_text(self.chat_id, message)
+        else:
+            for user_id in self.player_map.copy().keys():
+                await BotAPI.send_plain_text(user_id, message)
+
+    def build_timer(self, time, pending, label):
+        if self.chat_id == 0:
+            self.timer = PendingTimer(time, pending, label, self.player_map.copy().keys())
+        else:
+            self.timer = PendingTimer(time, pending, label, [self.chat_id])
 
     # добавить игрока в комнату
     async def add_member(self, user_data, is_admin=False):
@@ -35,7 +47,7 @@ class Room:
         self.player_map[user_id] = new_player
 
         print(user_id, '(admin = {0}) now is in room'.format(is_admin), self.room_id)
-        await self.send_plain_all('Игрок {0} зашёл в комнату!'.format(user_data.username))
+        await self.send_plain_all('Игрок {0} зашёл в комнату!'.format(user_data.username), send_to_group=True)
 
     # вывести список игроков
     async def list_member(self, user_data):
@@ -48,29 +60,50 @@ class Room:
 
         await BotAPI.send_plain_text(user_data.id, result)
 
+    # изменяем лимит по времени
+    async def reset_time(self, new_time):
+        self.time_for_round = new_time
+        await self.send_plain_all(
+            'Ведущий изменил лимит по времени! Новый лимит: {0} секунд'.format(new_time), send_to_group=True)
+
+    async def reset_quality(self, new_quality):
+        new_quality = max(new_quality, 1)
+        new_quality = min(new_quality, 10)
+        self.generation_quality = new_quality
+        await self.send_plain_all(
+            'Ведущий изменил качество генерации! Новое значение: {0}'.format(new_quality), send_to_group=True)
+
     # стереть комнату
-    def destroy_room(self):
+    async def destroy_room(self):
+        for user_id in self.player_map.copy().keys():
+            if user_id in self.player_map.keys():
+                await self.remove_member(user_id, destroy_on_empty=False)
         print('room {0} has been destroyed'.format(self.room_id))
         rooms.pop(self.room_id)
+        if self.chat_id != 0:
+            chats.pop(self.chat_id)
         self.timer = None
 
     # выкинуть игрока
-    async def remove_member(self, user_data):
-        user_id = user_data.id
+    async def remove_member(self, user_id, destroy_on_empty=True):
         is_admin = self.player_map[user_id].is_admin
+        username = self.player_map[user_id].username
         self.player_map.pop(user_id)
         players.pop(user_id)
 
         print(user_id, 'left the room ', self.room_id)
-        await self.send_plain_all('Игрок {0} покинул комнату!'.format(user_data.username))
+        await self.send_plain_all('Игрок {0} покинул комнату!'.format(username), send_to_group=True)
 
         # если в комнате пусто
-        if len(self.player_map) == 0:
-            self.destroy_room()
+        if len(self.player_map) == 0 and destroy_on_empty:
+            await self.destroy_room()
             return
 
         # если ливнул админ
-        if is_admin:
+        if is_admin and self.chat_id != 0 and destroy_on_empty:
+            await self.send_plain_all('Админ ливнул из игры!', send_to_group=True)
+            await self.destroy_room()
+        elif is_admin and destroy_on_empty:
             # ищем нового
             new_admin = None
             for user_id2 in self.player_map.copy().keys():
@@ -88,7 +121,9 @@ class Room:
         self.round = -1
         self.max_rounds = members
         self.tasks = []
-        self.timer = PendingTimer(11, [10, 5], 'Игра начнется через {0} секунд!', self.player_map.values())
+        self.build_timer(11,
+                         [10, 5],
+                         'Игра начнется через {0} секунд!')
         self.cycle_list = []
         self.cycle_map = {}
 
@@ -98,7 +133,6 @@ class Room:
             self.text_history[user_id] = []
 
         # генерим цикл
-        # TODO: проверить этот недокод :)
         for user_id in self.player_map.copy().keys():
             self.cycle_list.append(user_id)
         for iteration in range(members * 2):
@@ -119,15 +153,15 @@ class Room:
         all_users = all_users[:-2]
 
         # начинаем!
-        await self.send_plain_all('Ведущий запустил игру!')
-        await self.send_plain_all('Игроки: ' + all_users)
+        await self.send_plain_all('Ведущий запустил игру!', send_to_group=True)
+        await self.send_plain_all('Игроки: ' + all_users, send_to_group=True)
         self.timer.start()
 
     # стоп-игра
     async def stop_game(self):
         self.stage = 0
         self.timer = None
-        await self.send_plain_all('Игра окончена, ждём ведущего')
+        await self.send_plain_all('Игра окончена, ждём ведущего', send_to_group=True)
         return
 
     # принять текст от игрока
@@ -147,28 +181,27 @@ class Room:
     # начало текстового раунда (дана картинка - введи подпись/предложение)
     async def text_round_start(self):
         self.round += 1
-        self.timer = PendingTimer(30, [20, 10, 5], 'Осталось {0} секунд!', self.player_map.values())
+        time = self.time_for_round
+        self.build_timer(time, [time, time // 2, time // 5], 'Осталось {0} секунд!')
         self.timer.start()
         self.stage = 1
 
-        # добавляем пустую запись в историю (если чел вылетел - у него будет '' в качестве текста)
-        members = len(self.text_history)
+        # добавляем пустую запись в историю (если чел вылетел - у него будет '<ничего>' в качестве текста)
         for user_id in self.player_map.copy().keys():
-            self.text_history[user_id].append('')
+            self.text_history[user_id].append('<ничего>')
 
-    # начало кариночного раунда (нейросеть генерит картинки)
+    # начало картиночного раунда (нейросеть генерит картинки)
     async def picture_round_start(self):
         self.timer = None
         self.stage = 2
 
         # добавляем пустую запись в историю (когда нейронка сгенерит - тут будет путь к картинке/сама картинка)
-        members = len(self.image_history)
         for user_id in self.player_map.copy().keys():
-            self.image_history[user_id].append('')
-        #TODO: Отправляем тексты нейронке. Лучше всего нейронку пихать в отдельный поток
-        #TODO: т.к. иначе весь бот для ВСЕХ уйдёт в АФК при ожидании генерации
+            self.image_history[user_id].append('<ничего>')
+        # TODO: Отправляем тексты нейронке. Лучше всего нейронку пихать в отдельный поток
+        # TODO: т.к. иначе весь бот для ВСЕХ уйдёт в АФК при ожидании генерации
 
-    # апдейт кариночного раунда (нейросеть генерит картинки)
+    # апдейт картиночного раунда (нейросеть генерит картинки)
     async def picture_round_update(self):
         if len(self.tasks) == 0:
             # нейронка всё сгенерила!
@@ -190,7 +223,7 @@ class Room:
 
     # вывод всех альбомов
     async def print_history(self):
-        await self.send_plain_all('Альбомы: ')
+        await self.send_plain_all('Альбомы: ', send_to_group=True)
         for i in range(len(self.cycle_list)):
             # для каждой цепочки...
             user_id = self.cycle_list[i]
@@ -206,10 +239,8 @@ class Room:
                     # а это картинка!
                     result += self.image_history[user_id][j // 2]
                     result += ' ---> '
-
-
             result = result[:-6]  # удаляем '--->'
-            await self.send_plain_all(result)
+            await self.send_plain_all(result, send_to_group=True)
 
     # апдейт комнаты
     async def room_update(self):
@@ -244,8 +275,9 @@ async def global_update():
             # апдейтим комнату
             room = rooms[room_id]
             await room.room_update()
-        await asyncio.sleep(0.2)
+        await asyncio.sleep(0.5)
 
 
-rooms = {}     # 'ROOM_ID' --> Room, все комнаты
-players = {}   # 'TG_ID' --> 'ROOM_ID', какой игрок в какой комнате
+rooms = {}  # 'ROOM_ID' --> Room, все комнаты
+players = {}  # 'TG_ID' --> 'ROOM_ID', какой игрок в какой комнате
+chats = {}
